+++ cohere/client.py
@@ -2,6 +2,7 @@ import json as jsonlib
 import os
 from concurrent import futures
 from concurrent.futures import ThreadPoolExecutor
+from dataclasses import asdict
 from functools import partial
 from typing import Any, Dict, List, Optional, Union
 
@@ -10,7 +11,7 @@ from requests.adapters import HTTPAdapter
 from urllib3 import Retry
 
 import cohere
-from cohere.error import CohereAPIError, CohereError
+from cohere.error import CohereAPIError, CohereConnectionError, CohereError
 from cohere.logging import logger
 from cohere.responses import (
     Classification,
@@ -27,7 +28,11 @@ from cohere.responses.classify import LabelPrediction
 from cohere.responses.cluster import ClusterJobResult, CreateClusterJobResponse
 from cohere.responses.detectlang import DetectLanguageResponse, Language
 from cohere.responses.embeddings import Embeddings
-from cohere.responses.feedback import GenerateFeedbackResponse
+from cohere.responses.feedback import (
+    GenerateFeedbackResponse,
+    GeneratePreferenceFeedbackResponse,
+    PreferenceRating,
+)
 from cohere.responses.rerank import Reranking
 from cohere.responses.response import check_response
 from cohere.responses.summarize import SummarizeResponse
@@ -58,7 +63,7 @@ class Client:
         timeout: int = 120,
     ) -> None:
         self.api_key = api_key or os.getenv("CO_API_KEY")
-        self.api_url = cohere.COHERE_API_URL
+        self.api_url = os.getenv("CO_API_URL", cohere.COHERE_API_URL)
         self.batch_size = cohere.COHERE_EMBED_BATCH_SIZE
         self._executor = ThreadPoolExecutor(num_workers)
         self.num_workers = num_workers
@@ -151,15 +156,12 @@ class Client:
     def chat(
         self,
         query: str,
-        session_id: str = "",
         conversation_id: str = "",
         model: Optional[str] = None,
         return_chatlog: bool = False,
         return_prompt: bool = False,
         return_preamble: bool = False,
         chatlog_override: List[Dict[str, str]] = None,
-        persona_name: str = None,
-        persona_prompt: str = None,
         preamble_override: str = None,
         user_name: str = None,
         temperature: float = 0.8,
@@ -170,17 +172,14 @@ class Client:
 
         Args:
             query (str): The query to send to the chatbot.
-            session_id (str): Deprecated, use conversation_id instead.
             conversation_id (str): (Optional) The conversation id to continue the conversation.
             model (str): (Optional) The model to use for generating the next reply.
             return_chatlog (bool): (Optional) Whether to return the chatlog.
             return_prompt (bool): (Optional) Whether to return the prompt.
             return_preamble (bool): (Optional) Whether to return the preamble.
             chatlog_override (List[Dict[str, str]]): (Optional) A list of chatlog entries to override the chatlog.
-            persona_name (str): Deprecated.
-            persona_prompt (str): Deprecated, use preamble_override instead.
             preamble_override (str): (Optional) A string to override the preamble.
-            user_name (str): Deprecated.
+            user_name (str): (Optional) A string to override the username.
             temperature (float): (Optional) The temperature to use for the next reply. The higher the temperature, the more random the reply.
             max_tokens (int): (Optional) The max tokens generated for the next reply.
             stream (bool): Return streaming tokens.
@@ -190,7 +189,7 @@ class Client:
         Examples:
             A simple chat messsage:
                 >>> res = co.chat(query="Hey! How are you doing today?")
-                >>> print(res.reply)
+                >>> print(res.text)
                 >>> print(res.conversation_id)
             Continuing a session using a specific model:
                 >>> res = co.chat(
@@ -198,7 +197,7 @@ class Client:
                 >>>     conversation_id="1234",
                 >>>     model="command-xlarge",
                 >>>     return_chatlog=True)
-                >>> print(res.reply)
+                >>> print(res.text)
                 >>> print(res.chatlog)
             Overriding a chat log:
                 >>> res = co.chat(
@@ -210,7 +209,7 @@ class Client:
                 >>>         {'Bot': 'That is great to hear!'},
                 >>>     ],
                 >>>     return_chatlog=True)
-                >>> print(res.reply)
+                >>> print(res.text)
                 >>> print(res.chatlog)
             Streaming chat:
                 >>> res = co.chat(
@@ -221,25 +220,6 @@ class Client:
         """
         if chatlog_override is not None:
             self._validate_chatlog_override(chatlog_override)
-
-        if session_id != "":
-            conversation_id = session_id
-            logger.warning(
-                "The 'session_id' parameter is deprecated and will be removed in a future version of this function. Use 'conversation_id' instead.",
-            )
-        if persona_prompt is not None:
-            preamble_override = persona_prompt
-            logger.warning(
-                "The 'persona_prompt' parameter is deprecated and will be removed in a future version of this function. Use 'preamble_override' instead.",
-            )
-        if persona_name is not None:
-            logger.warning(
-                "The 'persona_name' parameter is deprecated and will be removed in a future version of this function.",
-            )
-        if user_name is not None:
-            logger.warning(
-                "The 'user_name' parameter is deprecated and will be removed in a future version of this function.",
-            )
 
         json_body = {
             "query": query,
@@ -253,6 +233,7 @@ class Client:
             "temperature": temperature,
             "max_tokens": max_tokens,
             "stream": stream,
+            "user_name": user_name,
         }
         response = self._request(cohere.CHAT_URL, json=json_body, stream=stream)
 
@@ -499,20 +480,63 @@ class Client:
         response = self._request(cohere.GENERATE_FEEDBACK_URL, json_body)
         return GenerateFeedbackResponse(id=response["id"])
 
+    def generate_preference_feedback(
+        self,
+        ratings: List[PreferenceRating],
+        model=None,
+        prompt: str = None,
+        annotator_id: str = None,
+    ) -> GeneratePreferenceFeedbackResponse:
+        """Give preference feedback on a response from the Cohere Generate API to improve the model.
+
+        Args:
+            ratings (List[PreferenceRating]): A list of PreferenceRating objects.
+            model (str): (Optional) ID of the model.
+            prompt (str): (Optional) The prompt used to generate the response.
+            annotator_id (str): (Optional) The ID of the annotator.
+
+        Examples:
+            A user accepts a model's suggestion in an assisted writing setting, and prefers it to a second suggestion:
+            >>> generations = co.generate(f"Write me a polite email responding to the one below: {email}. Response:", num_generations=2)
+            >>> if user_accepted_idx: // prompt user for which generation they prefer
+            >>>    ratings = []
+            >>>    if user_accepted_idx == 0:
+            >>>        ratings.append(PreferenceRating(request_id=0, rating=1))
+            >>>        ratings.append(PreferenceRating(request_id=1, rating=0))
+            >>>    else:
+            >>>        ratings.append(PreferenceRating(request_id=0, rating=0))
+            >>>        ratings.append(PreferenceRating(request_id=1, rating=1))
+            >>>    co.generate_preference_feedback(ratings=ratings)
+        """
+        ratings_dicts = []
+        for rating in ratings:
+            ratings_dicts.append(asdict(rating))
+
+        json_body = {
+            "ratings": ratings_dicts,
+            "prompt": prompt,
+            "annotator_id": annotator_id,
+            "model": model,
+        }
+        response = self._request(cohere.GENERATE_PREFERENCE_FEEDBACK_URL, json_body)
+        return GenerateFeedbackResponse(id=response["id"])
+
     def rerank(
         self,
         query: str,
         documents: Union[List[str], List[Dict[str, Any]]],
-        model: str = None,
+        model: str,
         top_n: Optional[int] = None,
+        max_chunks_per_doc: Optional[int] = None,
     ) -> Reranking:
         """Returns an ordered list of documents ordered by their relevance to the provided query
 
         Args:
             query (str): The search query
             documents (list[str], list[dict]): The documents to rerank
-            model (str): (Optional) The model to use for re-ranking
+            model (str): The model to use for re-ranking
             top_n (int): (optional) The number of results to return, defaults to returning all results
+            max_chunks_per_doc (int): (optional) The maximum number of chunks derived from a document
         """
         parsed_docs = []
         for doc in documents:
@@ -531,6 +555,7 @@ class Client:
             "model": model,
             "top_n": top_n,
             "return_documents": False,
+            "max_chunks_per_doc": max_chunks_per_doc,
         }
 
         reranking = Reranking(self._request(cohere.RERANK_URL, json=json_body))
@@ -580,16 +605,20 @@ class Client:
     def create_cluster_job(
         self,
         embeddings_url: str,
-        threshold: Optional[float] = None,
         min_cluster_size: Optional[int] = None,
+        n_neighbors: Optional[int] = None,
+        is_deterministic: Optional[bool] = None,
     ) -> CreateClusterJobResponse:
         """Create clustering job.
 
         Args:
             embeddings_url (str): File with embeddings to cluster.
-            threshold (Optional[float], optional): Similarity threshold above which two texts are deemed to belong in
-                the same cluster. Defaults to None.
-            min_cluster_size (Optional[int], optional): Minimum number of elements in a cluster. Defaults to None.
+            min_cluster_size (Optional[int], optional): Minimum number of elements in a cluster. Defaults to 10.
+            n_neighbors (Optional[int], optional): Number of nearest neighbors used by UMAP to establish the
+                local structure of the data. Defaults to 15. For more information, please refer to
+                https://umap-learn.readthedocs.io/en/latest/parameters.html#n-neighbors
+            is_deterministic (Optional[bool], optional): Determines whether the output of the cluster job is
+                deterministic. Defaults to True.
 
         Returns:
             CreateClusterJobResponse: Created clustering job handler
@@ -597,8 +626,9 @@ class Client:
 
         json_body = {
             "embeddings_url": embeddings_url,
-            "threshold": threshold,
             "min_cluster_size": min_cluster_size,
+            "n_neighbors": n_neighbors,
+            "is_deterministic": is_deterministic,
         }
 
         response = self._request(cohere.CLUSTER_JOBS_URL, json=json_body)