+++ cohere/responses/__init__.py
@@ -7,7 +7,11 @@ from cohere.responses.cluster import (
 )
 from cohere.responses.detectlang import DetectLanguageResponse, Language
 from cohere.responses.embeddings import Embeddings
-from cohere.responses.feedback import GenerateFeedbackResponse
+from cohere.responses.feedback import (
+    GenerateFeedbackResponse,
+    GeneratePreferenceFeedbackResponse,
+    PreferenceRating,
+)
 from cohere.responses.generation import Generation, Generations, StreamingGenerations
 from cohere.responses.rerank import RerankDocument, Reranking, RerankResult
 from cohere.responses.summarize import SummarizeResponse