+++ cohere/responses/feedback.py
@@ -1,22 +1,12 @@
-"""Give feedback on a response from the Cohere API to improve the model.
-
-Can be used programmatically like so:
-
-Example: a user accepts a model's suggestion in an assisted writing setting
-```
-generations = co.generate(f"Write me a polite email responding to the one below:\n{email}\n\nResponse:")
-if user_accepted_suggestion:
-    generations[0].feedback(good_response=True)
-```
-
-Example: the user edits the model's suggestion
-```
-generations = co.generate(f"Write me a polite email responding to the one below:\n{email}\n\nResponse:")
-if user_edits_suggestion:
-    generations[0].feedback(good_response=False, desired_response=user_edited_response)
-```
-"""
-
+from dataclasses import dataclass
 from typing import NamedTuple
 
 GenerateFeedbackResponse = NamedTuple("GenerateFeedbackResponse", [("id", str)])
+GeneratePreferenceFeedbackResponse = NamedTuple("GeneratePreferenceFeedbackResponse", [("id", str)])
+
+
+@dataclass
+class PreferenceRating:
+    request_id: str
+    rating: float
+    generation: str